Add default generic type to ArbitraryOrd
The `PartialOrd` trait uses a default generic type for the right hand
side of the comparison. This is useful for example to compare an object
to a foreign type. For the same reasons we should provide a default
generic type to `ArbitraryOrd`.

Add generic type to `ArbitraryOrd` and default to `Self`.

I believe this also makes the trait object safe.