Remove Eq trait bound
The `PartialOrd` trait is object safe but `Ord` is not. We can make
`ArbitraryOrd` object safe by softening the bounds and only requiring
`PartialEq` - we still provide a blanket implementation of `Ord` if `Eq`
is implemented for `T`.

Add a unit test that verifies using `Box<dyn T>` - props to Poelstra for
the idea.