Bump version to 1.0.0-alpha.0
In preparation for release bump the version number, add a changelog
entry, and update the lock files.

This is the alpha release. Explicitly includes deprecated code that will
be deleted in the `1.0` release.

This crate is the first of the so-called 'leaf crates' in
`rust-bitcoin`. Defined as crates we control at the bottom of the
dependency tree of the `bitcoin` crate. I know its supposed to be
dependency graph but then there would be no 'bottom'. There must be a
real term for this?